Move shared types to types.ts
I'm finding that having a dedicated file for types that are shared between
components makes it easier to avoid circular dependencies between components.

e.g. if a parent component requires a child component to be of a certain type
it's possible to get into a circular dependency if the type if declared in the
parent component file, imported by the child component which is in turn imported
by the parent component.