Improve the experience on mobile devices
The styling was completely broken on mobile. These changes make the layout
a little more reponsive (though there's still a long way to go before it's
a good experience on mobile - e.g. importing data should likely be a full
screen action instead of taking up 30+percent of the height).