Switch to using webpack
Something in the compilation pipeline for react-scripts was causing a function
to be undefined in the FNB-Default parser. It was found that  bank-schema-cli
wasn't seeing the same issues despite using the same library, which narrowed down
the issue to the transpilation of this project.

The switch has successfully resulted in the error going away. Took this opportunity
to add a bundle analyzer to the project.